+++ build.clj
@@ -3,7 +3,7 @@
             [deps-deploy.deps-deploy :as dd]))
 
 (def lib 'org.scicloj/zulipdata)
-(def version "0.1.0")
+(def version "0.2.0")
 (def snapshot (str version "-SNAPSHOT"))
 (def class-dir "target/classes")