+++ notebooks/zulipdata_book/views.clj
@@ -63,9 +63,8 @@
   ["clojurecivitas" "scicloj-webpublic" "gratitude" "events"])
 
 (def messages
-  (->> (pull/pull-channels! sample-channels)
-       (filter (fn [[k _]] (string? k)))
-       (mapcat (fn [[_ r]] (pull/all-messages r)))))
+  (-> (pull/pull-channels! sample-channels)
+      pull/all-channel-messages))
 
 (count messages)