+++ test/zulipdata_book/views_generated_test.clj
@@ -19,32 +19,29 @@
  v4_l65
  (def
   messages
-  (->>
-   (pull/pull-channels! sample-channels)
-   (filter (fn [[k _]] (string? k)))
-   (mapcat (fn [[_ r]] (pull/all-messages r))))))
+  (-> (pull/pull-channels! sample-channels) pull/all-channel-messages)))
 
 
-(def v5_l70 (count messages))
+(def v5_l69 (count messages))
 
 
-(def v7_l81 (def timeline (views/messages-timeline messages)))
+(def v7_l80 (def timeline (views/messages-timeline messages)))
 
 
-(def v8_l83 (tc/row-count timeline))
+(def v8_l82 (tc/row-count timeline))
 
 
-(deftest t9_l85 (is (= v8_l83 (count messages))))
+(deftest t9_l84 (is (= v8_l82 (count messages))))
 
 
-(def v11_l90 (-> timeline tc/column-names sort))
+(def v11_l89 (-> timeline tc/column-names sort))
 
 
 (deftest
- t12_l92
+ t12_l91
  (is
   (=
-   v11_l90
+   v11_l89
    '(:channel
      :client
      :content
@@ -60,32 +57,32 @@
      :timestamp))))
 
 
-(def v14_l100 (-> timeline (tc/order-by :instant :desc)))
+(def v14_l99 (-> timeline (tc/order-by :instant :desc)))
 
 
-(def v16_l107 (-> timeline :instant first type))
+(def v16_l106 (-> timeline :instant first type))
 
 
-(deftest t17_l109 (is (= v16_l107 java.time.Instant)))
+(deftest t17_l108 (is (= v16_l106 java.time.Instant)))
 
 
-(def v19_l115 (-> timeline (tc/select-rows :edited) tc/row-count))
+(def v19_l114 (-> timeline (tc/select-rows :edited) tc/row-count))
 
 
-(def v21_l123 (def reactions (views/reactions-long messages)))
+(def v21_l122 (def reactions (views/reactions-long messages)))
 
 
-(def v22_l125 reactions)
+(def v22_l124 reactions)
 
 
-(def v23_l127 (-> reactions tc/column-names sort))
+(def v23_l126 (-> reactions tc/column-names sort))
 
 
 (deftest
- t24_l129
+ t24_l128
  (is
   (=
-   v23_l127
+   v23_l126
    '(:channel
      :emoji-code
      :emoji-name
@@ -98,7 +95,7 @@
 
 
 (def
- v26_l135
+ v26_l134
  (->
   reactions
   (tc/group-by [:emoji-name])
@@ -107,20 +104,20 @@
   (tc/head 5)))
 
 
-(def v28_l150 (def edits (views/edits-long messages)))
+(def v28_l149 (def edits (views/edits-long messages)))
 
 
-(def v29_l152 edits)
+(def v29_l151 edits)
 
 
-(def v30_l154 (-> edits tc/column-names sort))
+(def v30_l153 (-> edits tc/column-names sort))
 
 
 (deftest
- t31_l156
+ t31_l155
  (is
   (=
-   v30_l154
+   v30_l153
    '(:channel
      :edit-ts
      :edit-user-id
@@ -131,23 +128,23 @@
      :stream-id))))
 
 
-(def v33_l166 (def links (views/topic-links-long messages)))
+(def v33_l165 (def links (views/topic-links-long messages)))
 
 
-(def v34_l168 links)
+(def v34_l167 links)
 
 
-(def v35_l170 (tc/column-names links))
+(def v35_l169 (tc/column-names links))
 
 
 (deftest
- t36_l172
+ t36_l171
  (is
-  (= v35_l170 [:message-id :stream-id :channel :link-text :link-url])))
+  (= v35_l169 [:message-id :stream-id :channel :link-text :link-url])))
 
 
 (def
- v38_l177
+ v38_l176
  (->
   links
   (tc/add-column
@@ -155,7 +152,7 @@
    (fn
     [ds]
     (mapv
-     (fn* [p1__51162#] (some-> p1__51162# (java.net.URI.) .getHost))
+     (fn* [p1__41842#] (some-> p1__41842# (java.net.URI.) .getHost))
      (:link-url ds))))
   (tc/group-by [:host])
   (tc/aggregate {:n tc/row-count})